\name{add.all.combinations}
\alias{add.all.combinations}
\title{Add all combinations}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add all combinations of the given rows and columns to the data frames.
}
\usage{add.all.combinations(data, vars = list(NULL, NULL))}
\arguments{
\item{data}{data.frame}
\item{vars}{variables (list of character vectors)}
}

\details{This function is used to ensure that we have a matrix of the appropriate
dimensionaliy with no missing cells.}

\examples{}
\keyword{internal}
