\name{add.all.combinations}
\alias{add.all.combinations}
\title{Add all combinations}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add all combinations of the given rows and columsn to the data frames.
}
\usage{add.all.combinations(data, rows, cols)}
\arguments{
\item{data}{data.frame}
\item{rows}{row variables (character vector)}
\item{cols}{column variables (character vector)}
}

\details{This function is used to ensure that we have a matrix of the appropriate
dimensionaliy with no missing cells.}

\examples{}
\keyword{manip}
