\name{as.data.frame.cast_df}
\alias{as.data.frame.cast_df}
\title{as.data.frame.cast\_df}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert cast data.frame into a matrix
}
\usage{\method{as.data.frame}{cast_df}(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{Strips off cast related attributes so data frame becomes a normal data frame}

\keyword{internal}
