\name{as.data.frame.cast_matrix}
\alias{as.data.frame.cast_matrix}
\title{Convert cast matrix into a data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{as.data.frame.cast_matrix(x, row.names, optional, ...)}
\arguments{
\item{x}{}
\item{row.names}{}
\item{optional}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
