\name{as.matrix.cast_df}
\alias{as.matrix.cast_df}
\title{as.matrix.cast\_df}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert cast data.frame into a matrix
}
\usage{as.matrix.cast_df(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{Converts a data frame produced by cast into a matrix with
appropriate dimnames.}

\keyword{internal}
