\name{as.matrix.cast_matrix}
\alias{as.matrix.cast_matrix}
\title{as.matrix.cast\_matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert cast matrix into a matrix
}
\usage{as.matrix.cast_matrix(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{Strips off cast related attributes so matrix becomes a normal matrix}

\examples{}
\keyword{internal}
