\name{as.data.frame.reshape_matrix}
\alias{as.data.frame.reshape_matrix}
\title{Convert reshape matrix into a data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{as.data.frame.reshape_matrix(x, row.names, optional)}
\arguments{
\item{x}{}
\item{row.names}{}
\item{optional}{}
}

\details{}

\examples{}
\keyword{manip}
