\name{cast_matrix}
\alias{cast_matrix}
\title{Cast matrix.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Createa a new cast matrix
}
\usage{cast_matrix(m, dimnames)}
\arguments{
\item{m}{matrix to turn into cast matrix}
\item{dimnames}{list of dimension names (as data.frames), row, col, ...}
}
\value{object of type \code{\link{cast_matrix}}}
\details{}

\examples{}
\keyword{internal}
