\name{check_formula}
\alias{check_formula}
\title{Check formula}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Checks that formula is a valid reshaping formula.
}
\usage{check_formula(formula, varnames)}
\arguments{
\item{formula}{formula to check}
\item{varnames}{vector of variable names}
}

\details{}

\examples{}
\keyword{internal}
