\name{clean.vars}
\alias{clean.vars}
\title{Clean variables.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Clean variable list for reshaping.
}
\usage{clean.vars(vars)}
\arguments{
\item{vars}{vector of variable names}
}
\value{Vector of "real" variable names (excluding result\_variable etc.)}

\keyword{internal}
