\name{colsplit}
\alias{colsplit}
\alias{colsplit.factor}
\alias{colsplit.character}
\title{Split a vector into multiple columns}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function can be used to split up a column that has been pasted together.
}
\usage{colsplit(x, split="", names)}
\arguments{
\item{x}{character vector or factor to split up}
\item{split}{regular expression to split on}
\item{names}{names for output columns}
}

\details{}

\examples{}
\keyword{manip}
