\name{compactify}
\alias{compactify}
\title{Compactify}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compacts an array into a smaller array of lists containing the remaining dimensions
}
\usage{compactify(x, dims = 1)}
\arguments{
\item{x}{array}
\item{dims}{number of dimension to compact}
}

\details{}

\examples{}
\keyword{internal}
