\name{compute.margins}
\alias{compute.margins}
\title{Compute margins}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute marginal values.
}
\usage{compute.margins(data, margins, vars, fun.aggregate, ..., df=FALSE)}
\arguments{
\item{data}{data frame}
\item{margins}{margins to compute}
\item{vars}{all id variables}
\item{fun.aggregate}{aggregation function}
\item{...}{other argument passed to aggregation function}
\item{df}{}
}

\details{}

\examples{}
\keyword{internal}
