\name{compute.margins}
\alias{compute.margins}
\title{Compute margins}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute marginal values.
}
\usage{compute.margins(data, margins, fun.aggregate, ...)}
\arguments{
\item{data}{data frame}
\item{margins}{margins to compute}
\item{fun.aggregate}{aggregation function}
\item{...}{other argument passed to aggregation function}
}

\details{}

\examples{}
\keyword{manip}
