\name{condense.df}
\alias{condense.df}
\title{Condense a data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Condense
}
\usage{condense.df(data, variables, fun, ...)}
\arguments{
\item{data}{data frame}
\item{variables}{character vector of variables to condense over}
\item{fun}{function to condense with}
\item{...}{arguments passed to condensing function}
}


\keyword{manip}
