\name{dim_names}
\alias{dim_names}
\title{Dimension names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience method for extracting row and column names
}
\usage{dim_names(data, vars)}
\arguments{
\item{data}{data frame}
\item{vars}{variables to use}
}

\details{}

\examples{}
\keyword{internal}
