\name{expand.grid.df}
\alias{expand.grid.df}
\title{expand grid}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
expand grid of data frames
}
\usage{expand.grid.df(df1, df2, unique=TRUE)}
\arguments{
\item{df1}{data frame 1 (varies fastest)}
\item{df2}{data frame 2}
\item{unique}{}
}

\details{Creates new data frame containing all combination of rows from df1 and df2.}

\examples{}
\keyword{manip}
