\name{expand.grid.df}
\alias{expand.grid.df}
\title{Expand grid}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Expand grid of data frames
}
\usage{expand.grid.df(df1, df2)}
\arguments{
\item{df1}{data frame 1}
\item{df2}{data frame 2}
}

\details{Creates new data frame containing all combination of rows from df1 and df2.}

\examples{}
\keyword{manip}
