\name{French fries}
\docType{data}
\alias{french_fries}
\title{Sensory data from a french fries experiment}
\description{
This data was collected from a sensory experiment conducted at Iowa State University in 2004.  The investigators were interested in the effect of using three different fryer oils had on the taste of the fries.

Variables:

\itemize{
\item time in weeks from start of study.
\item treatment (type of oil), 
\item subject, 
\item replicate, 
\item potato-y flavour, 
\item buttery flavour, 
\item grassy flavour,
\item rancid flavour,
\item painty flavour 
}
}
\usage{data(french_fries)}
\format{A data frame with 696 rows and 9 variables}
\references{}
\keyword{datasets}
