\name{"[.cast_df"}
\alias{[.cast_df}
\title{Subset a cast data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Subset a cast data frame. This will result in a normal dataframe, because it is
}
\usage{"[.cast_df"(x, i=1:nrow(x), j=1:ncol(x), ..., drop = FALSE)}
\arguments{
\item{x}{cast\_df object to subset}
\item{i}{row indices}
\item{j}{column indices}
\item{...}{other arguments}
\item{drop}{discard extra dimensions?}
}

\details{}

\examples{}
\keyword{internal}
