\name{"[.cast_matrix"}
\alias{[.cast_matrix}
\title{Subset a cast matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Subset a cast matrix just like you would a normal matrix.  Preserves row and column names nicely.
}
\usage{"[.cast_matrix"(x, i=1:nrow(x), j=1:ncol(x), ..., drop=FALSE)}
\arguments{
\item{x}{cast matrix object}
\item{i}{row indices}
\item{j}{col indices}
\item{...}{other arguments not used}
\item{drop}{discard extra dimensions?}
}

\details{}

\examples{}
\keyword{internal}
