\name{melt}
\alias{melt}
\title{Melt}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Melt an object into a form suitable for easy casting.
}
\usage{melt(data, ...)}
\arguments{
\item{data}{Data set to melt}
\item{...}{Other arguments passed to the melt function}
}

\details{See \code{\link{melt.data.frame}} and \code{\link{melt.array}} for details.}

\examples{}
\keyword{manip}
