\name{melt.cast_df}
\alias{melt.cast_df}
\title{Melt cast data.frames}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
After casting into a particular form, it can sometimes be useful to
}
\usage{melt.cast_df(data, drop.margins=TRUE, ...)}
\arguments{
\item{data}{}
\item{drop.margins}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
