\name{melt.cast_matrix}
\alias{melt.cast_matrix}
\title{Melt cast matrices}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Melt the results of a cast
}
\usage{\method{melt}{cast_matrix}(data, ...)}
\arguments{
\item{data}{}
\item{...}{other arguments ignored}
}

\details{Converts to a data frame and then uses \code{\link{melt.cast_df}}}

\keyword{internal}
