\name{merge_all}
\alias{merge_all}
\title{Merge together a series of data.frames}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{merge_all(dfs, ...)}
\arguments{
\item{dfs}{list of data frames to merge}
\item{...}{}
}

\details{Order of data frames should be from most complete to least complete}
\seealso{\code{\link{merge_recurse}}}
\examples{}
\keyword{manip}
