\name{merge_recurse}
\alias{merge_recurse}
\title{Recursively merge data frames}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{merge_recurse(dfs, ...)}
\arguments{
\item{dfs}{list of data frames to merge}
\item{...}{}
}

\details{}
\seealso{\code{\link{merge_all}}}
\examples{}
\keyword{internal}
