\name{nested.by}
\alias{nested.by}
\title{Nested.by function}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Nest series of by statements returning nested list
}
\usage{nested.by(data, INDICES, FUN, ...)}
\arguments{
\item{data}{}
\item{INDICES}{}
\item{FUN}{}
\item{...}{}
}

\details{Work horse for producing cast lists.}

\examples{}
\keyword{internal}
