\name{nulldefault}
\alias{nulldefault}
\title{Null default}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Use default value when null
}
\usage{nulldefault(x, default)}
\arguments{
\item{x}{}
\item{default}{}
}

\details{Handy method when argument defaults aren't good enough.}

\keyword{internal}
