\name{parse_formula}
\alias{parse_formula}
\title{Parse formula}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Parse an R formula into convenient chucks
}
\usage{parse_formula(formula, varnames)}
\arguments{
\item{formula}{formula to parse}
\item{varnames}{variable names}
}

\details{}

\examples{}
\keyword{manip}
