\name{prettyprint}
\alias{prettyprint}
\title{Print reshaped data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{prettyprint(x, digits=getOption("digits"), ..., colnames=TRUE)}
\arguments{
\item{x}{}
\item{digits}{}
\item{...}{}
\item{colnames}{}
}

\details{}

\examples{}
\keyword{internal}
