\name{print.cast_df}
\alias{print.cast_df}
\title{Print reshaped data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{print.cast_df(x, digits=getOption("digits"), ..., colnames=TRUE)}
\arguments{
\item{x}{}
\item{digits}{}
\item{...}{}
\item{colnames}{}
}

\details{}

\examples{}
\keyword{internal}
