\name{print.reshape_df}
\alias{print.reshape_df}
\title{Print reshaped data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{print.reshape_df(x, digits=getOption("digits"), ...)}
\arguments{
\item{x}{}
\item{digits}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{manip}
