\name{print.reshape_matrix}
\alias{print.reshape_matrix}
\title{Print reshaped matrix.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{print.reshape_matrix(x, digits=getOption("digits"), ...)}
\arguments{
\item{x}{reshaped matrix}
\item{digits}{number of digits to show}
\item{...}{other arguments to match generic}
}

\details{}

\examples{}
\keyword{manip}
