\name{rdimnames}
\alias{rdimnames}
\alias{rdimnames<-}
\alias{rcolnames}
\alias{rcolnames<-}
\alias{rrownames}
\alias{rrownames<-}
\title{Dimension names}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
These methods provide easy access to the special dimension names
}
\usage{rdimnames(x)}
\arguments{
\item{x}{}
}

\details{}

\examples{}
\keyword{internal}
