\name{rescaler.data.frame}
\alias{rescaler.data.frame}
\title{Rescale a data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Rescales data frame by columns
}
\usage{rescaler.data.frame(x, type="sd", ...)}
\arguments{
\item{x}{data.frame to rescale}
\item{type}{type of rescaling to apply}
\item{...}{other arguments passed to rescaler}
}

\details{}

\examples{}
\keyword{internal}
