\name{rescaler.default}
\alias{rescaler.default}
\title{Default rescaler}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See \code{\link{rescaler}} for details
}
\usage{\method{rescaler}{default}(x, type="sd", ...)}
\arguments{
\item{x}{vector to rescale}
\item{type}{type of rescaling to apply}
\item{...}{other arguments passed to rescaler}
}


\keyword{internal}
