\name{rescaler.matrix}
\alias{rescaler.matrix}
\title{Rescale a matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Rescales matrix by columns
}
\usage{\method{rescaler}{matrix}(x, type="sd", ...)}
\arguments{
\item{x}{matrix to rescale}
\item{type}{type of rescaling to apply}
\item{...}{other arguments passed to rescaler}
}


\keyword{internal}
