\name{round_any}
\alias{round_any}
\title{Round any}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Round to multiple of any number
}
\usage{round_any(x, round, f=floor)}
\arguments{
\item{x}{numeric vector to round}
\item{round}{number to round to}
\item{f}{function to use for round (eg. \code{\link{floor}})}
}

\details{}

\examples{}
\keyword{internal}
