% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{smiths}
\alias{smiths}
\title{Demo data describing the Smiths.}
\format{A data frame with 2 rows and 5 variables}
\usage{
smiths
}
\description{
A small demo dataset describing John and Mary Smith.  Used in the
introductory vignette.
}
\keyword{datasets}

