\name{sort.df}
\alias{sort.df}
\title{Sort data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience method for sorting a data frame using the given variables.
}
\usage{sort.df(data, vars=names(data))}
\arguments{
\item{data}{data frame to sort}
\item{vars}{variables to use for sorting}
}

\details{}

\examples{}
\keyword{manip}
