\name{str.cast_matrix}
\alias{str.cast_matrix}
\alias{str.cast_df}
\alias{print.cast_matrix}
\alias{print.cast_df}
\title{Print cast objects}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Printing methods
}
\usage{str.cast_matrix(object, ...)}
\arguments{
\item{object}{}
\item{...}{}
}

\details{Used for printing.}

\keyword{internal}
