\name{strip.dups}
\alias{strip.dups}
\title{Strip duplicates.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Strips out duplicates from data.frame and replace them with periods.
}
\usage{strip.dups(df)}
\arguments{
\item{df}{data.frame to modify}
}
\value{character matrix}
\details{}

\examples{}
\keyword{internal}
