\name{uniquedefault}
\alias{uniquedefault}
\title{Unique default}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience function for setting default if not unique
}
\usage{uniquedefault(values, default)}
\arguments{
\item{values}{vector of values}
\item{default}{default to use if values not uniquez}
}

\details{}

\examples{}
\keyword{manip}
