\name{as.data.frame.cast_matrix}
\alias{as.data.frame.cast_matrix}
\title{as.data.frame.cast\_matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert cast matrix into a data frame
}
\usage{\method{as.data.frame}{cast_matrix}(x, row.names, optional, ...)}
\arguments{
\item{x}{}
\item{row.names}{}
\item{optional}{}
\item{...}{}
}

\details{Converts a matrix produced by cast into a data frame with
appropriate id columns.}

\keyword{internal}
