\name{cast_parse_formula}
\alias{cast_parse_formula}
\title{cast parse formula}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Parse formula for casting
}
\usage{cast_parse_formula(formula = "...  ~ variable", varnames)}
\arguments{
\item{formula}{}
\item{varnames}{}
}
\value{
 \item{row character vector of row names}
 \item{col character vector of column names}
 \item{aggregate boolean whether aggregation will occur}
}
\details{}

\examples{}
\keyword{internal}
