\name{melt.cast_df}
\alias{melt.cast_df}
\title{Melt cast data.frames}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Melt the results of a cast
}
\usage{melt.cast_df(data, drop.margins=TRUE, ...)}
\arguments{
\item{data}{}
\item{drop.margins}{}
\item{...}{}
}

\details{This can be useful when performning complex aggregations - melting
the result of a cast will do it's best to figure out the correct variables
to use as id and measured.}

\examples{}
\keyword{internal}
