% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.r
\name{melt.default}
\alias{melt.default}
\title{Melt a vector.
For vectors, makes a column of a data frame}
\usage{
\method{melt}{default}(data, ..., na.rm = FALSE, value.name = "value")
}
\arguments{
\item{data}{vector to melt}

\item{...}{further arguments passed to or from other methods.}

\item{na.rm}{Should NA values be removed from the data set? This will
convert explicit missings to implicit missings.}

\item{value.name}{name of variable used to store values}
}
\description{
Melt a vector.
For vectors, makes a column of a data frame
}
\seealso{
\code{\link{melt}}, \code{\link{cast}}

Other melt methods: \code{\link{melt.array}},
  \code{\link{melt.data.frame}}, \code{\link{melt.list}}
}
\keyword{manip}

