\name{prettyprint}
\alias{prettyprint}
\title{Pretty print}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print reshaped data frame
}
\usage{prettyprint(x, digits=getOption("digits"), ..., colnames=TRUE)}
\arguments{
\item{x}{}
\item{digits}{}
\item{...}{}
\item{colnames}{}
}

\details{This will always work on the direct output from cast, but may not
if you have manipulated (e.g. subsetted) the results.}

\examples{}
\keyword{internal}
