% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recast}
\alias{recast}
\title{Recast: melt and cast in a single step}
\usage{
recast(data, formula, ..., id.var, measure.var)
}
\arguments{
\item{data}{data set to melt}

\item{formula}{casting formula, see \code{\link{dcast}} for specifics}

\item{...}{other arguments passed to \code{\link{dcast}}}

\item{id.var}{identifying variables. If blank, will use all non
measure.var variables}

\item{measure.var}{measured variables. If blank, will use all non
id.var variables}
}
\description{
This conveniently wraps melting and (d)casting a data frame into
a single step.
}
\examples{
recast(french_fries, time ~ variable, id.var = 1:4)
}
\seealso{
\url{http://had.co.nz/reshape/}
}
\keyword{manip}

