\name{rescaler.matrix}
\alias{rescaler.matrix}
\title{Rescale a matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{rescaler.matrix(x, type="sd", ...)}
\arguments{
\item{x}{matrix to rescale}
\item{type}{type of rescaling to apply}
\item{...}{other arguments passed to rescaler}
}

\details{}

\examples{}
\keyword{internal}
