\name{str.cast_matrix}
\alias{str.cast_matrix}
\alias{str.cast_df}
\alias{print.cast_matrix}
\alias{print.cast_df}
\title{Printing methods}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{str.cast_matrix(object, ...)}
\arguments{
\item{object}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
