\name{updatelist}
\alias{updatelist}
\title{Update list}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Update a list, but don't create new entries
}
\usage{updatelist(x, y)}
\arguments{
\item{x}{list to be updated}
\item{y}{list with updated values}
}

\details{}

\examples{}
\keyword{internal}
